Fix README
DONT DO THIS

Do not think in having a 'compound history'

It looks ugly in the log/network view.

The squash is WAY better. Just do not use `-m` flag
with commit, because git will give a nice commit
message with all commits squashed in the commit editor.
Just add a title like 'Import repo/branch' (or tag)

BTW: The tags in this repository will have a hard-coded
namespace. I don't care GitHub tells it is a invalid name.

PS: GitHub do NOT detect duplicated commits in different
repositories.